move fetch_mcp function to amplicon_utils, write unit test for assess_mcp_proportions
