+++ mgnify_pipelines_toolkit/analysis/amplicon/assess_inflection_point_mcp.py
@@ -5,7 +5,7 @@ from Bio.Seq import Seq
 import numpy as np
 import pandas as pd
 
-from assess_mcp_proportions import fetch_mcp
+from mgnify_pipelines_toolkit.analysis.amplicon.amplicon_utils import fetch_mcp
 from mgnify_pipelines_toolkit.analysis.amplicon.amplicon_utils import get_read_count, build_cons_seq, build_mcp_cons_dict_list
 
 def parse_args():
